# Conflicts:
#	source/actors/Player.hx

+++ source/actors/Actor.hx
@@ -1,6 +1,5 @@
 package actors;
 
-import flixel.FlxG;
 import flixel.FlxObject;
 import flixel.FlxSprite;
 import flixel.group.FlxGroup.FlxTypedGroup;
@@ -19,11 +18,12 @@ class Actor extends FlxSprite {
 	public var goesLeft:Bool;
 	public var goesRight:Bool;
 	public var mA:Float;
-	public var selectedWeapon:BaseItem;
+	public var indexOfselectedWeapon:Int;
 	public var isPlayer:Bool;
-	public var bullets:FlxTypedGroup<Projectile>;
 
-	var playState:PlayState;
+	// Buffable properties
+	public static inline var SPEED = "speed";
+	public static inline var HEALTH = "health";
 
 	public function new(?X:Float = 0, ?Y:Float = 0) {
 		super(X, Y);
@@ -44,17 +44,16 @@ class Actor extends FlxSprite {
 
 	override public function update(elapsed:Float):Void {
 		super.update(elapsed);
-		playState = cast FlxG.state;
-		FlxG.collide(this, playState._mWalls);
-
 	}
 
 	public static function takeDamage(p:Projectile, a:Actor):Void {
-		p.kill();
-		a.hurt(p.damage);
+		if ((p.owner.isPlayer && !a.isPlayer) || (!p.owner.isPlayer && a.isPlayer)) {
+			p.kill();
+			a.hurt(p.damage);
 
-		if (a.health <= 0) {
-			a.healthBar.kill();
+			if (a.health <= 0) {
+				a.healthBar.kill();
+			}
 		}
 	}
 }
